[event_engine] Listener file descriptors

+++ src/core/lib/event_engine/posix_engine/posix_endpoint.cc
@@ -486,11 +486,12 @@ void PosixEndpointImpl::UpdateRcvLowat() {
   // TODO(ctiller): Check if supported by OS.
   // TODO(ctiller): Allow some adjustments instead of hardcoding things.
 
-  static constexpr int kRcvLowatMax = 16 * 1024 * 1024;
+  static constexpr uint32_t kRcvLowatMax = 16 * 1024 * 1024;
   static constexpr int kRcvLowatThreshold = 16 * 1024;
 
-  int remaining = std::min({static_cast<int>(incoming_buffer_->Length()),
-                            kRcvLowatMax, min_progress_size_});
+  uint32_t remaining =
+      std::min({static_cast<uint32_t>(incoming_buffer_->Length()), kRcvLowatMax,
+                min_progress_size_});
 
   // Setting SO_RCVLOWAT for small quantities does not save on CPU.
   if (remaining < kRcvLowatThreshold) {
@@ -610,7 +611,8 @@ bool PosixEndpointImpl::Read(absl::AnyInvocable<void(absl::Status)> on_read,
   incoming_buffer_->Clear();
   incoming_buffer_->Swap(last_read_buffer_);
   if (args != nullptr && grpc_core::IsTcpFrameSizeTuningEnabled()) {
-    min_progress_size_ = std::max(static_cast<int>(args->read_hint_bytes), 1);
+    min_progress_size_ = std::max(static_cast<uint32_t>(args->read_hint_bytes),
+                                  static_cast<uint32_t>(1));
   } else {
     min_progress_size_ = 1;
   }
